test_automation/requirements.txt
# Python Test Automation Requirements

# Appium for Android automation
Appium-Python-Client>=3.1.0

# Data processing
pandas>=2.0.0

# Testing framework (built-in unittest is used, but pytest is useful too)
pytest>=7.4.0

# Additional utilities
python-dateutil>=2.8.2
